Second simulation method file dump
The original method to simulate the titration, opening an instance of python to run a script, is clunky and hard to work with. This file dump contains the files needed to perform a different method of simulating.

The executable titration_simulator.exe, whose source can be found in titration_simulator_source.py, reads from the file called "simulation_parameters.txt", parses the data inside the file, performs the simulation, and outputs the volume and pH data to another file called "simulation_results.txt". 

All that the C# would need to do is set the parameters, run the executable, and read and plot the results.